Using JDBC for our database.
In this new project we are using JDBC instead of the libraries we were using earlier for our database.
A database class is added to interact with the database. We are still following the MVC architecture/approach.
For now we have a couple of classes as our models, we are still to add more. There is only one controller that handles all the requests from the UI.
Navigation to a home page added :).